+++ examples/02_typescript/src/DisplayRemoteData.tsx
@@ -17,16 +17,9 @@ const DisplayRemoteData: React.FC<{ id: string }> = ({ id }) => {
   const url = `https://jsonplaceholder.typicode.com/posts/${id}`;
   const asyncTask = useAsyncTaskFetch<{ title: string }>(url);
   useAsyncRun(asyncTask);
-  const {
-    pending,
-    error,
-    result,
-    abort,
-  } = asyncTask;
-  if (error) return <Err error={error} />;
-  if (pending) return <Loading abort={abort} />;
-  if (!result) return <div>No result</div>;
-  return <div>RemoteData: {result.title}</div>;
+  if (asyncTask.error) return <Err error={asyncTask.error} />;
+  if (asyncTask.pending) return <Loading abort={asyncTask.abort} />;
+  return <div>RemoteData: {asyncTask.result.title}</div>;
 };
 
 export default DisplayRemoteData;