+++ examples/03_startbutton/src/DisplayRemoteData.tsx
@@ -16,19 +16,10 @@ const Loading: React.SFC<{ abort: () => void }> = ({ abort }) => (
 const DisplayRemoteData: React.FC<{ id: string }> = ({ id }) => {
   const url = `https://jsonplaceholder.typicode.com/posts/${id}`;
   const asyncTask = useAsyncTaskFetch<{ title: string }>(url);
-  const {
-    started,
-    pending,
-    error,
-    result,
-    start,
-    abort,
-  } = asyncTask;
-  if (!started) return <button type="button" onClick={() => start()}>start</button>;
-  if (error) return <Err error={error} />;
-  if (pending) return <Loading abort={abort} />;
-  if (!result) return <div>No result</div>;
-  return <div>RemoteData: {result.title}</div>;
+  if (!asyncTask.started) return <button type="button" onClick={() => asyncTask.start()}>start</button>;
+  if (asyncTask.error) return <Err error={asyncTask.error} />;
+  if (asyncTask.pending) return <Loading abort={asyncTask.abort} />;
+  return <div>RemoteData: {asyncTask.result.title}</div>;
 };
 
 export default DisplayRemoteData;